Tagging Release 1.2.7.

+++ dmtcp/src/util_init.cpp
@@ -19,13 +19,18 @@
  *  <http://www.gnu.org/licenses/>.                                         *
  ****************************************************************************/
 
+#include <stdlib.h>
 #include <string.h>
+#include <string>
+#include <sstream>
+#include <errno.h>
+#include <dlfcn.h>
 #include "constants.h"
-#include "util.h"
+#include  "util.h"
+#include  "uniquepid.h"
+#include  "../jalib/jassert.h"
+#include  "../jalib/jfilesystem.h"
 #include "protectedfds.h"
-#include "uniquepid.h"
-#include "../jalib/jassert.h"
-#include "../jalib/jfilesystem.h"
 
 void dmtcp::Util::initializeLogFile(dmtcp::string procname, dmtcp::string prevLogPath)
 {
@@ -55,7 +60,7 @@ void dmtcp::Util::initializeLogFile(dmtcp::string procname, dmtcp::string prevLo
 
   a << "\nArgv: ";
   dmtcp::vector<dmtcp::string> args = jalib::Filesystem::GetProgramArgs();
-  size_t i;
+  int i;
   for (i = 0; i < args.size(); i++) {
     a << " " << args[i];
   }
